+++ spatialprofilingtoolbox/apiserver/Dockerfile.append
@@ -1,3 +1,2 @@
-RUN pip3 install "$WHEEL_FILENAME[countsserver]"
 COPY ./app ./
 CMD ["uvicorn", "app.main:app", "--host", "0.0.0.0", "--port", "8080", "--log-level", "debug", "--workers", "9"]
